+++ src/utils/menu.py
@@ -1,0 +1,17 @@
+
+# This is the main menu 
+
+main_menu = """
+                Main Menu
+                0 - Exit 
+                1 - Product Menu
+                """
+
+product_menu = """
+                Product Menu
+                0 - Return To Main Menu
+                1 - List Products 
+                2 - Create New Products 
+                3 - Update Products
+                4 - Delete Products
+                """
